Fix removal of numeric suffixes in labels
